requirements.txt
streamlit
replicate
pdfminer.six

